apache-pulsar/requirements.txt
pulsar-client==3.1.0
